pyproject.toml
[tool.poetry]
name = "huma-underwriter-simple-creditline"
version = "0.1.0"
description = ""
authors = ["Ji Peng <ji@huma.finance>", "Jiatu Liu <jiatu@huma.finance>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
structlog = "^22.3.0"
fastapi = "^0.95.1"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"
httpx = "^0.24.0"
orjson = "^3.8.10"
huma-signals = "^0.2.1"
# TODO(jiatu): remove the pin after the dependent packages fix this issue.
#  urllib3 version 2.0 and above breaks a bunch of dependent packages.
urllib3 = "^1.26"
datadog = "^0.45.0"
ddtrace = {extras = ["opentracing"], version = "^1.13.3"}
pydantic = "^1.10.8"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.0"
black = "^22.10.0"
pytest = "^7.2.0"
pytest-describe = "^2.0.1"
pre-commit = "^3.1.0"
mypy = "^1.0.1"
isort = "^5.10.1"
flake8 = "^6.0.0"
pylint = "^2.16.2"
pylint-google-style-guide-imports-enforcing = "^1.3.0"
pylint-pydantic = "^0.1.6"
pytest-spec = "^3.2.0"
vcrpy = "^4.2.1"
pytest-asyncio = "^0.21.0"
factory-boy = "^3.2.1"
types-factory-boy = "^0.4.1"

[tool.pytest.ini_options]
describe_prefixes = ["describe", "if", "when", "with", "without"]
python_functions = ["it_*", "its_*"]
asyncio_mode = "auto"

[tool.autoflake]
in-place = true
remove-all-unused-imports = true
ignore-init-module-imports = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
plugins = ["pydantic.mypy"]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = ["vcr"]
ignore_missing_imports = true

[tool.pylint.messages_control]
# C0103, C0114, C0115, C0116: docstring and file name formats.
# R0801: similar lines.
# R0903: too few public methods.
# W0511: TODOs.
disable = "C0103, C0114, C0115, C0116, R0801, R0903, W0511"
# https://github.com/samuelcolvin/pydantic/issues/1961
extension-pkg-whitelist = "pydantic, orjson"

[tool.pylint.format]
max-line-length = "120"

[tool.pylint.master]
load-plugins = "pylint_google_style_guide_imports_enforcing, pylint_pydantic"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
